requirements.txt
openaq
pandas
python-dotenv
sqlalchemy
psycopg2-binary
python-dateutil